(#66) Allow nodes to fail during task::wait playbook
It could be that nodes are restarting, or something about the task
interrupting service or maybe the task is trying to do recovery
after some bad deploy - regardless there are cases where we just can
not assume all nodes will respond on this status call.

Now when someone passes fail_ok it signals that they wish to ignore
failures while waiting but could lead to surprising outcomes where
1 out of 1000 nodes did not fail and 999 failed - and then the wait
will finish, use with caution.