(#113) Ensure APT repository is managed before package
The module does not enforce the choria repository to be managed before
the choria package. When adding this module to a node, the first run may
fail because Puppet attempts to install the choria package which is not
yet available.

Force the choria package to be managed after `apt update` has been run
(which wil happen after the repo is added) to fix the issue.