(#111) Disable repo_gpgcheck for yum repos
Disables the GPG check of yum repo metadata for choria yum release and
nightly repos.

This is a change to make it easier to set up internal mirrors. Note
that this also disables the check for people not using a mirror, but
this should be fully compatible.

There is a tradeoff in security here as users will no longer have the
benefit of GPG signed metadata. In future we may want to make this
behaviour opt-in.