(#147) Added ssldir parameter to broker config.
Setting ssldir independently for choria broker and choria server has a different net effect from setting this value in /etc/choria/plugins.d/choria.conf. This setting was added so it can be overridden in both.

This is important in situations where the puppetserver is installed in a nonstandard manner, for example when using puppetserver in containers, or with pupperware project.