Add a bolt task to setup shell completion
This is a building block that might be used when/if we want to add Bolt
tasks / plans to setup Choria on nodes not managed by Puppet, in
particular client nodes.

Currently our systems administrator manually add the choria repo and
install the software, but this is not enough to have shell completion
and we might need to update it from time to time.

Add a bolt task that make managing the completions easier:

```
bolt task run choria::completion -t localhost --run-as root
```