(#320) allow plugin.choria.network.gateway_name to be set
Recent NATS Server versions require unique cluster names
for leafnodes connecting to a cluster.

We default to CHORIA for all cluster names, this used to
work but failed in the previous choria release.

Those wishing to use leafnodes to a Choria cluster would
need to set this variable to a unique name

Signed-off-by: R.I.Pienaar <rip@devco.net>