(#326) Update stdlib requirements
In order to avoid function name collision, the puppet community is
moving away from unscoped functions in modules so that only Puppet
provide functions without an explicit scope.

stdlib 9.0.0 deprecate all unscoped Puppet 4 functions and provide them
with the `stdlib::` scope.  While the unscoped function still exist and
default to calling the scoped function, Puppet 8 enabled strict mode by
default in an effort to limit the accumulation of cruft in the end users
control-repos.  That mean that if we enable the latest stdlib, we must
also increase the minimum version to make sure the scoped versions of
the functions are available.